requirements.txt
albumentations==1.3.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
coverage==7.4.3
cycler==0.12.1
exceptiongroup==1.2.0
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.47.2
fsspec==2024.3.1
idna==3.6
imageio==2.33.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.11.2
MarkupSafe==2.1.5
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
pluggy==1.3.0
pyparsing==3.1.1
pytest==7.4.4
pytest-cov==4.1.0
python-dateutil==2.8.2
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
ruff==0.1.14
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tifffile==2024.1.30
tomli==2.0.1
torch==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.9.0
urllib3==2.2.1
pydicom==2.4.4
Pillow==10.2.0
Werkzeug==3.0.2
ultralytics==8.1.45